Some tweaks to make it more macOS compliant
On macOS, we use `Preferences`, not `Options`. On Linux and Windows, I've seen both used, so I guess we can just prefer `Preferences`.

`Launch at Login` is a more commonly used term in macOS apps.

I also removed `label` from the Quit menus as it's handled for us when `role: 'quit'` is defined.